requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
distro==1.9.0
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
nvdlib==0.7.7
oauthlib==3.2.2
openai==1.14.2
psycopg2-binary==2.9.10
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==72.1.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.41
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.1.3
wheel==0.44.0
